Fix: 선택된 항목이 없을 때 alert 표시 보완

+++ src/main/resources/static/js/home.js
@@ -1,19 +1,25 @@
 // 체크박스 클릭 후, 모두 삭제 버튼
 document.getElementById('delete-button').addEventListener('click', () => {
+    const tbody = document
+
     const selectKeys = document.querySelectorAll('input[name=selectedKey]:checked');
     const selectkeyList = Array.from(selectKeys).map(selectKeys => selectKeys.value);
+    const selectkeyLength = selectkeyList.length;
 
-    const promises = selectkeyList.map(key =>
-        fetch(`/view/products/${key}`, {
-                    method: 'DELETE'
-                })
-    );
+    if (!selectkeyLength) {
+        alert('선택된 물품이 없습니다!')
+    }
+    else {
+        const promises = selectkeyList.map(key =>
+            fetch(`/view/products/${key}`, {
+                        method: 'DELETE'
+                    })
+        );
 
-    Promise.all(promises)
-        .then(() => {
-            alert('선택 물품 삭제 완료!');
-            window.location.reload(); // 새로고침
-        });
-
-
+        Promise.all(promises)
+            .then(() => {
+                alert('선택 물품 삭제 완료!');
+                window.location.reload(); // 새로고침
+            });
+    }
 });
